move logic & improve code
